% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted_values_clognorm.R
\name{fitted_values_clognorm}
\alias{fitted_values_clognorm}
\title{Generate fitted values from a estimated GAM (censored log normal version)}
\usage{
fitted_values_clognorm(
  object,
  data = NULL,
  scale = c("response", "link", "linear predictor"),
  ci_level = 0.95,
  ...
)
}
\arguments{
\item{object}{a fitted model using \code{clognorm} as the response.}

\item{data}{optional data frame of covariate values for which fitted values
are to be returned.}

\item{scale}{character; what scale should the fitted values be returned on?
use \code{"response"} to obtain intervals on the response scale (exponentiated)
for a log-normally distributed response.}

\item{ci_level}{numeric; a value between 0 and 1 indicating the coverage of
the credible interval.}

\item{...}{arguments passed to \code{\link[mgcv:predict.gam]{mgcv::predict.gam()}}. Note that \code{type},
\code{newdata}, and \code{se.fit} are already used and passed on to
\code{\link[mgcv:predict.gam]{mgcv::predict.gam()}}.}
}
\value{
A tibble (data frame) whose first \emph{m} columns contain either the data
used to fit the model (if \code{data} was \code{NULL}), or the variables supplied to
\code{data}. Four further columns are added:
\itemize{
\item \code{fitted}: the fitted values on the specified scale,
\item \code{se}: the standard error of the fitted values (always on the \emph{link} scale),
\item \code{lower}, \code{upper}: the limits of the credible interval on the fitted values,
on the specified scale.
}
}
\description{
This is a stand-in replacement for \code{gratia::fitted_values} to be used with
the clognorm family and make predictions on the response scale.
}
\seealso{
clognorm
}
