% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_family_cdf.R
\name{fix_family_cdf}
\alias{fix_family_cdf}
\title{Fix $cdf() slot of a mgcv family}
\usage{
fix_family_cdf(fam)
}
\arguments{
\item{fam}{an \code{mgcv} family}
}
\description{
Fix $cdf() slot of a mgcv family.
}
\examples{
library(mgcvUtils); library(MASS)
b <- gam(list(accel~s(times,k=20,bs="ad"),~s(times)),
         data=mcycle,family=gaulss())

b$family <- fix_family_cdf( b$family )

# Quantile residuals should look uniformly distributed
qres <- b$family$cdf(mcycle$accel, b$fitted.values,
                     wt = rep(1, nrow(mcycle)), scale = 1)
hist(qres)
# Not too bad
}
