% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_family_qf.R
\name{fix_family_qf}
\alias{fix_family_qf}
\title{Fix $qf() slot of a mgcv family}
\usage{
fix_family_qf(fam)
}
\arguments{
\item{fam}{an \code{mgcv} family}
}
\description{
Fix $qf() slot of a mgcv family.
}
\examples{
library(mgcv);library(MASS)
b <- gam(list(accel~s(times,k=20,bs="ad"),~s(times)),
         data=mcycle,family=gaulss())

b$family <- fix_family_qf( b$family )
qfit <- sapply(c(0.1, 0.5, 0.9),
               function(.q)
                 b$family$qf(.q, b$fitted.values))

plot(mcycle, ylim = c(-150, 100))
for(ii in 1:ncol(qfit)){
  lines(mcycle$times, qfit[ , ii], col = 2)
}

}
