% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_family_rd.R
\name{fix_family_rd}
\alias{fix_family_rd}
\title{Fix $rd() slot of a mgcv family}
\usage{
fix_family_rd(fam)
}
\arguments{
\item{fam}{an \code{mgcv} family}
}
\description{
Fix $rd() slot of a mgcv family.
}
\examples{
library(gamUtils)
set.seed(525)
dat <- gamSim(1,n=400,dist="normal",scale=2)
b <- gam(list(y~s(x0),~s(x1),~1),data=dat,family=gevlss)

head( simulate(b, nsim = 2) )

}
