% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcvUtils-package.R
\docType{package}
\name{mgcvUtils}
\alias{mgcvUtils}
\title{Utilities for mgcv}
\description{
Here are some utilities to use with \code{mgcv}. This page will be populated with themes to help users get around.
}
