% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_exp.R
\name{obs_exp}
\alias{obs_exp}
\title{Observed versus expected diagnostic}
\usage{
obs_exp(model, covar, cut = NULL)
}
\arguments{
\item{model}{a fitted GAM}

\item{covar}{covariate to aggregate by (\code{character})}

\item{cut}{vector of cut points to aggregate at. If not supplied, the unique
values of \code{covar} are used.}
}
\value{
\code{data.frame} with values of observed and expected counts.
}
\description{
Given a covariate, calculate the observed and expected counts for each
unique value of the covariate. This can be a useful goodness of fit check
for GAMs when modelling count data. It might make sense in other cases.
}
\details{
One strategy for model checking is to calculate observed and expected counts
at different aggregations of the variable. If these match well then the
model fit is good.
}
\examples{
# simulate poisson data...
set.seed(6)
dat <- gamSim(1,n=200,dist="poisson",scale=.1)

b2<-gam(y~s(x0)+s(x1)+s(x2)+
        s(x3),family=poisson,data=dat,method="REML", control=gam.control(keepData=TRUE))
plot(b2,pages=1)
}
\author{
David L Miller, on the suggestion of Mark Bravington.
}
