% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adaptive.R
\name{plot_adaptive_1d}
\alias{plot_adaptive_1d}
\title{Plot univariate adaptive smoother penalties}
\usage{
plot_adaptive_1d(model, term = 1, bs = "ps", m = 2, nplot = 100)
}
\arguments{
\item{model}{a fitted \code{mgcv} model}

\item{term}{which term number in the model we want to plot}

\item{bs}{the basis used for the smoothing parameter smoother}

\item{m}{the order of the smoothing parameter smoother}

\item{nplot}{number of points used to make the grid for the plot}
}
\value{
just makes a plot, invisably returns the data underlying said plot
}
\description{
If we use an adaptive smoother, we can plot how the smoothing parameter
changes as a function of the covariate.
}
\examples{
library(mgcv)
library(MASS)

# fit an adaptive smooth to the mcycle data
b_ad <- gam(accel ~ s(times, bs="ad"), data=mcycle)

par(mfrow=c(1,2))
plot(b_ad)
plot_adaptive_1d(b_ad)
}
\author{
David L Miller with idea from Philip Dixon
}
