% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adaptive.R
\name{plot_adaptive_2d}
\alias{plot_adaptive_2d}
\title{Plot bivariate adaptive smoother penalties}
\usage{
plot_adaptive_2d(model, term = 1, bs = "ps", m = 2, nplot2 = 100)
}
\arguments{
\item{model}{a fitted \code{mgcv} model}

\item{term}{which term number in the model we want to plot}

\item{bs}{the basis used for the smoothing parameter smoother}

\item{m}{the order of the smoothing parameter smoother}

\item{nplot2}{number of points used to make the grid for the plot}
}
\value{
just makes a plot and invisibly returns the data used to make said
plot
}
\description{
If we use an adaptive smoother, we can plot $\lambda(x,y)$ over $x,y$ to
give an idea of how the smoothing parameter changes as a function of the
covariate.
}
\details{
This function does not cover all corner cases, your mileage may vary!
}
\examples{
# 2D example from ?smooth.construct.ad.smooth.spec

par(mfrow=c(2,2))
x <- seq(-.5, 1.5, length= 60)
z <- x
f3 <- function(x,z,k=15) { r<-sqrt(x^2+z^2);f<-exp(-r^2*k);f}
f <- outer(x, z, f3)
op <- par(bg = "white")

## Plot truth....
persp(x,z,f,theta=30,phi=30,col="lightblue",ticktype="detailed")

n <- 2000
x <- runif(n)*2-.5
z <- runif(n)*2-.5
f <- f3(x,z)
y <- f + rnorm(n)*.1

## Try tprs for comparison...
b0 <- gam(y~s(x, z, k=150))
vis.gam(b0, c("x", "z"), plot.type="contour")

## Now adaptive...
b <- gam(y~s(x, z, bs="ad", k=15, m=5), gamma=1.4)
vis.gam(b, c("x", "z"), plot.type="contour")

plot_adaptive_2d(b)
}
\author{
David L Miller with idea from Philip Dixon
}
