% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_with_ci.R
\name{pred_with_ci}
\alias{pred_with_ci}
\title{Make a prediction and put a confidence interval around it}
\usage{
pred_with_ci(
  model,
  data,
  alpha = 0.05,
  label = NULL,
  label_label = NULL,
  pred_label = "pred",
  ci_label = "ci",
  ci_types = c("u", "l"),
  sep = "_"
)
}
\arguments{
\item{model}{a fitted GAM}

\item{data}{the data we wish to predict to}

\item{alpha}{the level we want our confidence intervals at}

\item{label}{label for these models (e.g., "model_1")}

\item{label_label}{the column name for the model labels}

\item{pred_label}{the column name for the predictions}

\item{ci_label}{start of the column name for the confidence intervals}

\item{ci_types}{a 2-vector with the suffixes for the confidence intervals
(upper then lower)}

\item{sep}{separator between \code{ci_label} and \code{ci_types} in the confidence
intervals column name}
}
\description{
Small utility function to append a \code{data.frame} with predictions from a GAM
and include a confidence interval. Various things are customizable.
}
\details{
By the \code{data} argument (a \code{data.frame}) will be returned with 3 extra
columns, one for the prediction (column name as in \code{pred_label}), then upper
CI (named \code{paste(ci_label, ci_types[1], sep="_")}) and finally the
lower CI (named \code{paste(ci_label, ci_types[2], sep="_")}). If \code{label}
and \code{label_label} are non-empty, a column will be added (with name
\code{label_label}) and contents as in \code{label}.
}
\examples{
library(mgcvUtils)

# fit something to look at
set.seed(2)
dat <- gamSim(1,n=400,dist="normal",scale=2)
b <- gam(y~s(x0),data=dat)

# adding prediction/ci columns
dat2 <- pred_with_ci(b, dat)
head(dat2)

# add labels, so you can concatenate predictions later for
# plotting with ggplot2
dat3 <- pred_with_ci(b, dat, label="model1", label_label="model_id")
head(dat3)
}
\author{
David L Miller
}
