% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqgam_check.R
\name{rqgam_check}
\alias{rqgam_check}
\title{Randomised quantile residuals check plot}
\usage{
rqgam_check(gam.obj, ...)
}
\arguments{
\item{gam.obj}{a \code{\link[mgcv:gam]{gam}}, \code{\link[stats:glm]{glm}}}

\item{...}{arguments passed on to all plotting functions}
}
\value{
just plots!
}
\description{
Reproduces the "Resids vs. linear pred" plot from
\code{\link[mgcv:gam.check]{gam.check}} but using randomised quantile residuals, a la
Dunn and Smyth (1996). Check for heteroskedasticity as as usual, looking
for "funnel"-type structures in the points, which is much easier with
randomised quantile residuals than with deviance residuals, especially when
your model uses a count distribution as the response.
}
\details{
Note that this function only works with negative binomial and Tweedie
response distributions.
}
\author{
Based on code by Natalie Kelly, bugs added by Dave Miller
}
