% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_concurvity.R
\name{vis_concurvity}
\alias{vis_concurvity}
\title{Visualise concurvity between terms in a GAM}
\usage{
vis_concurvity(model, type = "estimate")
}
\arguments{
\item{model}{fitted model}

\item{type}{concurvity measure to plot, see \code{\link[mgcv:concurvity]{concurvity}}}
}
\description{
Plot measures of how much one term in the model could be explained by
another. When values are high, one should consider re-running variable
selection with one of the offending variables removed to check for stability
in term selection.
}
\details{
These methods are considered somewhat experimental at this time. Consult
\code{\link[mgcv:concurvity]{concurvity}} for more information on how concurvity
measures are calculated.
}
\examples{
# example from ?concurvity
library(mgcv)
## simulate data with concurvity...
set.seed(8);n<- 200
f2 <- function(x) 0.2 * x^11 * (10 * (1 - x))^6 + 10 *
            (10 * x)^3 * (1 - x)^10
t <- sort(runif(n)) ## first covariate
## make covariate x a smooth function of t + noise...
x <- f2(t) + rnorm(n)*3
## simulate response dependent on t and x...
y <- sin(4*pi*t) + exp(x/20) + rnorm(n)*.3

## fit model...
b <- gam(y ~ s(t,k=15) + s(x,k=15),method="REML")

## assess concurvity between each term and `rest of model'...
vis_concurvity(b)
}
\author{
David L Miller
}
