% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_nei.R
\name{vis_nei}
\alias{vis_nei}
\title{Visualize NCV neighbourhoods}
\usage{
vis_nei(nei)
}
\arguments{
\item{nei}{a \code{list} of elements as required by \code{mgcv::gam}}
}
\value{
\code{ggplot2} plot
}
\description{
Create a little plot showing what the neighbourhoods look like. The
resulting plot has number of rows equal to the number of folds and number of
columns equal to the number of data. Each tile is coloured according to
whether it's in the set of dropped indices, prediction points or both.
}
\author{
David L Miller
}
