% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogn.R
\name{clognorm}
\alias{clognorm}
\title{Censored log-normal distribution for \code{mgcv}}
\usage{
clognorm(theta = NULL, link = "identity", base = 10)
}
\arguments{
\item{theta}{kept as a placeholder, but only estimated \code{theta} is supported.}

\item{link}{The link function: only '"identity"' is supported}

\item{base}{the base of the logarithm used. base may be any number or "e"}
}
\value{
\code{family} object
}
\description{
Implementation of a censored log-normal distribution. This uses \code{mgcv}'s
built-in \code{cnorm} distribution but includes extra features, such as
transformation of the response (according to the specified \code{base}) and a
special \code{predict} function to provide unbiased predicted values using the
arithmetic mean. Note that this merely fits using \code{cnorm} and includes
post-hoc transformations to give "correct" results. See [\link[mgcv:cnorm]{mgcv::cnorm}] for
more information on how to specify censoring.
}
\examples{
library(mgcv)
library(mgcvUtils)
library(ggplot2)

# make a 1D example with less noise
set.seed(42)

# simulated example from ?gam
dat <- gamSim(1, n=200, dist="normal", scale=1)

e <- rnorm(nrow(dat), sd=2)
dat$y2 <- 10^(dat$f2/10 + e)

# censor values below some value
dat$y2 <- cbind(dat$y2, dat$y2)
# indicate which values are censored, see ?cnorm for rules
dat$y2[dat$y2[, 1] < 0.01, 2] <- -Inf
dat$y2[dat$y2[, 1] < 0.01, 1] <- 0.01

# response will be transformed by clognorm before fitting
b2 <- gam(y2~s(x2), data=dat, method="REML", family=clognorm(base=10))
summary(b2)

# compare predictions (on a random subset of the data)
pred <- dat[sample(1:nrow(dat), 100), ]
pred$y2 <- NULL
# uncorrected
old <- predict(b2, newdata=pred, se=TRUE)
# corrected (must use type="response")
new <- predict(b2, newdata=pred, type="response", se=TRUE)

preddat <- data.frame(pred.old = 10^old$fit,
                      pred.cor = new$fit,
                      pred.upper.old = 10^(old$fit + 1.96*old$se.fit),
                      pred.upper.cor = 10^(log10(new$fit) + 1.96*new$se.fit),
                      pred.lower.old = 10^(old$fit - 1.96*old$se.fit),
                      pred.lower.cor = 10^(log10(new$fit) - 1.96*new$se.fit))

# compare by plot
ggplot(preddat) +
  geom_point(aes(x=pred.old, y=pred.cor)) +
  geom_segment(aes(x=pred.lower.old, xend=pred.upper.old,
                   y=pred.cor, yend=pred.cor)) +
  geom_segment(aes(y=pred.lower.cor, yend=pred.upper.cor,
                   x=pred.old, xend=pred.old)) +
  geom_abline(intercept=0, slope=1, colour="red", lty=2) +
  labs(x="Uncorrected predictions", y="Corrected predictions") +
  theme_minimal()
}
\author{
David L Miller with code modified from mgcv by Simon Wood
}
