% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogn.R
\name{clognorm}
\alias{clognorm}
\title{Censored log-normal distribution for \code{mgcv}}
\usage{
clognorm(theta = NULL, link = "identity", base = 10)
}
\arguments{
\item{theta}{log standard deviation parameter. If supplied and positive then
taken as a fixed value of standard deviation (not its log). If supplied and
negative taken as negative of initial value for standard deviation (not its
log).}

\item{link}{The link function: '"identity"', '"log"' or '"sqrt"'.}

\item{base}{the base of the logarithm used. base may be any number or "e"}
}
\value{
\code{family} object
}
\description{
Censored log-normal distribution for \code{mgcv}
}
\examples{
library(mgcv)
library(mgcvUtils)
library(ggplot2)

# simulated example from ?gam
dat <- gamSim(1,n=2000,dist="normal",scale=1)

# make a 1D example with less noise
set.seed(2)

e <- rnorm(nrow(dat), sd=2)
dat$y2 <- exp(dat$f2/10 + e)

# censor values below some value
dat$y2 <- cbind(dat$y2, dat$y2)
dat$y2[dat$y2[, 1] < 0.01, 2] <- -Inf
dat$y2[dat$y2[, 1] < 0.01, 1] <- 1

b2 <- gam(log(y2)~s(x2),data=dat, method="REML", family=clognorm(base=10))
summary(b2)

# compare predictions (on a random subset of the data)
pred <- dat[sample(1:nrow(dat), 100), ]
pred$y2 <- NULL
old <- (predict(b2, newdata=pred, se=TRUE))
# corrected
new <- predict(b2, newdata=pred, type="response", se=TRUE)

preddat <- data.frame(pred.old = 10^old$fit,
                      pred.cor = new$fit,
                      pred.upper.old = 10^(old$fit + 1.96*old$se.fit),
                      pred.upper.cor = 10^(log10(new$fit) + 1.96*new$se.fit),
                      pred.lower.old = 10^(old$fit - 1.96*old$se.fit),
                      pred.lower.cor = 10^(log10(new$fit) - 1.96*new$se.fit))

ggplot(preddat) +
  geom_point(aes(x=pred.old, y=pred.cor)) +
  geom_segment(aes(x=pred.lower.old, xend=pred.upper.old, y=pred.cor, yend=pred.cor)) +
  geom_segment(aes(y=pred.lower.cor, yend=pred.upper.cor, x=pred.old, xend=pred.old)) +
  theme_minimal()
}
\author{
David L Miller with code modified from mgcv by Simon Wood
}
